Initial commit
First pass at the script. It simply takes keys from the dict and concatenates them together
to form a password based on how many and how long you want the words to be.

`words_dictionary.json` is where the words are coming from.

Added basic information to the `README.md` concerning usage and expected output.