Update to CLI
This patch updates the command line parameters to use argparse.
Argparse simplifies the use of the command line arguments and also provides an automatic -h help parameter.

Pyperclip is also used to copy the password directly to clipboard instead of printing to the console in the event that the private flag is set.