Added random flag to args and refactored password generation
Added the random flag to the args for when a user wants a random password. When using the -r flag the [
word and letter flags are now ignored. -p can also be used with -r.

Changed the way user determined password length works. only the number has to be added to the args. letter
and word lenghts no longer need to be assigned by saying -w=4 -l=4 can simple be 4 4.

Refactored how args are instantiated by moving them to main and not as global variables. Now handling the
making of the word list once rather than every loop iteration (that was dumb lol).

Added a little more info to the output when a a password is generated.