Manager and Docs update
This commit updates the manager to allow the handling of loose entries (see docs) and handle loading and saving of vault files.

The docs have been updated to list the functions of the class in a clear and easy to read form and also explain certain design decisions.

Fixes made to layout to adhere to PEP8.